Gracefully fail on race condition DB insert
Will handle and log an exception if a race condition occurs where a
duplicate value is inserting into the DB, i.e. when concurrent workers
attempt to insert a record with the same name.  Also allows uuid name
generation when a name is already taken via the
'ckanext.harvest.dup_append_type' config key, which greatly reduces the
chances of the above scenario occurring due to not using sequentially
numbered data records.